+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - ADDED: #3763 Generated Columns support added.
 - ADDED: #3471 Window Functions support added.
 - ADDED: #3470 UPDATE FROM clause support added.
+- ADDED: #3470 VACUUM INTO clause support added.
 - CHANGE: Windows binaries are now 64-bit.
 - CHANGE: SQLite2 plugin removed due to harder and harder maintenance of SQLite2 library compilation.
 - CHANGE: Minumum Qt version now is 5.12.0.